LayoutTests/http/tests/loading/cached-scripts-expected.txt

main frame - didStartProvisionalLoadForFrame
main frame - willCloseFrame
main frame - didCommitLoadForFrame
main frame - didClearWindowObjectForFrame
frame "<!--framePath //<!--frame0-->-->" - didClearWindowObjectForFrame
frame "<!--framePath //<!--frame0-->-->" - didStartProvisionalLoadForFrame
http://127.0.0.1:8000/loading/resources/cached-scripts-subframe.html - willSendRequest <NSURLRequest http://127.0.0.1:8000/loading/resources/cached-scripts-subframe.html> redirectResponse (null)
<unknown> - willCacheResponse: called
main frame - didFinishDocumentLoadForFrame
<unknown> - didFinishLoading
http://127.0.0.1:8000/loading/resources/cached-scripts-subframe.html - willSendRequest <NSURLRequest http://127.0.0.1:8000/loading/resources/cached-scripts-subframe.html> redirectResponse (null)
http://127.0.0.1:8000/loading/resources/cached-scripts-subframe.html - didReceiveResponse <NSURLResponse http://127.0.0.1:8000/loading/resources/cached-scripts-subframe.html>
frame "<!--framePath //<!--frame0-->-->" - willCloseFrame
frame "<!--framePath //<!--frame0-->-->" - didCommitLoadForFrame
frame "<!--framePath //<!--frame0-->-->" - didClearWindowObjectForFrame
http://127.0.0.1:8000/loading/resources/cached-scripts-subframe.html - willCacheResponse: called
data:text/javascript, - willSendRequest <NSURLRequest data:text/javascript,> redirectResponse (null)
http://127.0.0.1:8000/loading/resources/cached-scripts-subframe.html - didFinishLoading
data:text/javascript, - didReceiveResponse <NSURLResponse data:text/javascript,>
data:text/javascript,%20window.ranScript%20=%20true - willSendRequest <NSURLRequest data:text/javascript,%20window.ranScript%20=%20true> redirectResponse (null)
data:text/javascript, - didFinishLoading
data:text/javascript,%20window.ranScript%20=%20true - didReceiveResponse <NSURLResponse data:text/javascript,%20window.ranScript%20=%20true>
frame "<!--framePath //<!--frame0-->-->" - didFinishDocumentLoadForFrame
frame "<!--framePath //<!--frame0-->-->" - willPerformClientRedirectToURL: about:blank 
frame "<!--framePath //<!--frame0-->-->" - didHandleOnloadEventsForFrame
main frame - didHandleOnloadEventsForFrame
frame "<!--framePath //<!--frame0-->-->" - didFinishLoadForFrame
main frame - didFinishLoadForFrame
data:text/javascript,%20window.ranScript%20=%20true - didFinishLoading
frame "<!--framePath //<!--frame0-->-->" - didStartProvisionalLoadForFrame
about:blank - willSendRequest <NSURLRequest about:blank> redirectResponse (null)
about:blank - didReceiveResponse <NSURLResponse about:blank>
frame "<!--framePath //<!--frame0-->-->" - willCloseFrame
frame "<!--framePath //<!--frame0-->-->" - didCancelClientRedirectForFrame
frame "<!--framePath //<!--frame0-->-->" - didCommitLoadForFrame
frame "<!--framePath //<!--frame0-->-->" - didClearWindowObjectForFrame
frame "<!--framePath //<!--frame0-->-->" - willPerformClientRedirectToURL: http://127.0.0.1:8000/loading/resources/cached-scripts-subframe.html 
frame "<!--framePath //<!--frame0-->-->" - didHandleOnloadEventsForFrame
frame "<!--framePath //<!--frame0-->-->" - didFinishDocumentLoadForFrame
frame "<!--framePath //<!--frame0-->-->" - didFinishLoadForFrame
about:blank - didFinishLoading
frame "<!--framePath //<!--frame0-->-->" - didStartProvisionalLoadForFrame
http://127.0.0.1:8000/loading/resources/cached-scripts-subframe.html - willSendRequest <NSURLRequest http://127.0.0.1:8000/loading/resources/cached-scripts-subframe.html> redirectResponse (null)
http://127.0.0.1:8000/loading/resources/cached-scripts-subframe.html - willSendRequest <NSURLRequest http://127.0.0.1:8000/loading/resources/cached-scripts-subframe.html> redirectResponse (null)
http://127.0.0.1:8000/loading/resources/cached-scripts-subframe.html - didReceiveResponse <NSURLResponse http://127.0.0.1:8000/loading/resources/cached-scripts-subframe.html>
frame "<!--framePath //<!--frame0-->-->" - willCloseFrame
frame "<!--framePath //<!--frame0-->-->" - didCancelClientRedirectForFrame
frame "<!--framePath //<!--frame0-->-->" - didCommitLoadForFrame
frame "<!--framePath //<!--frame0-->-->" - didClearWindowObjectForFrame
frame "<!--framePath //<!--frame0-->-->" - didHandleOnloadEventsForFrame
frame "<!--framePath //<!--frame0-->-->" - didFinishDocumentLoadForFrame
frame "<!--framePath //<!--frame0-->-->" - didFinishLoadForFrame

This tests that when script that calls document.write to write out two new script tags while being parsed causes both scripts to be executed, even if the first one is cached.
SUCCESS
